Modifies arguments in main.py and splits constants.py after code review
