+++ utilities/env_constants.py
@@ -32,59 +32,7 @@ NUM_OF_ROUTES_PROP = os.environ.get("NUM_OF_ROUTES_PROP")
 ROUTE_TYPE_PROP = os.environ.get("ROUTE_TYPE_PROP")
 CATALOG_PROP = os.environ.get("CATALOG_PROP")
 
-# Define regex pattern for dataset version entity code in response retrieved by SPARQL query
-SPARQL_ENTITY_CODE_REGEX = "/(Q.+?)-"
-
 # External services
 SPARQL_URL = os.environ.get("SPARQL_URL")
 SPARQL_BIGDATA_URL = os.environ.get("SPARQL_BIGDATA_URL")
 API_URL = os.environ.get("API_URL")
-
-STAGING_SPARQL_URL = (
-    "http://staging.mobilitydatabase.org:8282//proxy/wdqs/bigdata/namespace/wdq/sparql"
-)
-STAGING_SPARQL_BIGDATA_URL = "http://staging.mobilitydatabase.org:8989/bigdata/sparql"
-STAGING_API_URL = "http://staging.mobilitydatabase.org/w/api.php"
-PRODUCTION_SPARQL_URL = (
-    "http://mobilitydatabase.org:8282//proxy/wdqs/bigdata/namespace/wdq/sparql"
-)
-PRODUCTION_SPARQL_BIGDATA_URL = "http://mobilitydatabase.org:8989/bigdata/sparql"
-PRODUCTION_API_URL = "http://mobilitydatabase.org/w/api.php"
-
-ID = "id"
-CLAIMS = "claims"
-MAINSNAK = "mainsnak"
-DATAVALUE = "datavalue"
-VALUE = "value"
-RESULTS = "results"
-BINDINGS = "bindings"
-LABELS = "labels"
-ENGLISH = "en"
-DATATYPE = "datatype"
-RANK = "rank"
-PROP_ID = "prop_id"
-IF_EXISTS = "if_exists"
-NORMAL = "normal"
-PREFERRED = "preferred"
-GLOBE_PRECISION = 1.0e-6
-GLOBE_URL = "http://www.wikidata.org/entity/Q2"
-LAT = "latitude"
-LON = "longitude"
-
-STOP_LAT = "stop_lat"
-STOP_LON = "stop_lon"
-IS_ADDITION = "is_addition"
-IS_MAXIMUM = "is_maximum"
-
-STOP_KEY = "stop"
-STATION_KEY = "station"
-ENTRANCE_KEY = "entrance"
-STOP = 0
-STATION = 1
-ENTRANCE = 2
-LOCATION_TYPE = "location_type"
-
-SPARQL_A = "a"
-SVC_PROP_URL_PATH = "/prop/statement/"
-SVC_ENTITY_URL_PATH = "/entity/"
-SVC_URL = "http://wikibase.svc"
